Renames and refactors quote-fetching component
Converts from fetching multiple quotes to a single quote,
streamlining the data transformation process. Updates file
names and exports for consistency. Adjusts UI to display a
single quote, aiding clarity and reducing complexity.
