+++ src/components/index.ts
@@ -1,6 +1,6 @@
 export * from './CodePen';
 export * from './ExternalLink';
-export * from './GetQuotes';
+export * from './GetQuote';
 export * from './Header';
 export * from './ImageHeader';
 export * from './Layout';
